Fold the unrouted request tests into an existing test file
Every entry in the CI test matrix uses one of the credentials files in
secrets.tar.gpg, and there are exactly as many of those files as there are
entries, so a new entry has no database to use and its job fails while
copying the file. Move the tests into tests/mock_vws/test_invalid_given_id.py,
which already covers requests which name something the API does not serve,
rather than adding an entry.
